requirements.txt
annotated-types==0.6.0
asttokens==2.4.1
backcall==0.2.0
blinker==1.7.0
Brotli==1.1.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
dnspython==2.6.1
exceptiongroup==1.1.3
executing==2.0.0
Flask==3.0.2
Flask-Compress==1.14
Flask-Cors==4.0.0
gunicorn==21.2.0
ipykernel==6.26.0
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.5.0
jupyter_core==5.4.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
pure-eval==0.2.2
pyaml==23.12.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.16.1
pymongo==4.6.2
python-dateutil==2.8.2
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
six==1.16.0
stack-data==0.6.3
tornado==6.3.3
traitlets==5.12.0
typing_extensions==4.10.0
wcwidth==0.2.8
Werkzeug==3.0.1
